refactor: staticmethod to utils
